memstate: Use /proc/<pid>/smaps_rollup instead of /proc/<pid>/smaps
Reading smaps file to print process memory usage info (PSS) is quite
time consuming - not to mention CPU intensive. On UEK5 and newer kernels,
there is a new smaps_rollup file in procfs, which is pretty much a summary
of the various memory stats across all mappings for that process, which is
what we care about. Reading smaps_rollup instead of smaps results in a
drastic increase in performance. The con is that this file is not available
in older kernels (UEK4 and below) - for those kernels, `memstate -p` will
not print a summary of all process' PSS values.

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Jose Lombera <jose.lombera@oracle.com>