Add scanfs subcommand
scanfs: Scan KVM images for corruption. Supports XFS and EXT4.

This tool allows us to scan XFS and EXT4 filesystems for corruption.
It takes reflink copies of the backup images and then scans these
reflinks for corruption.

Converted the old shell script to python.

This tool takes the path to backup image files as an argument. And
provides options to do a setup and cleanup.

Orabug: 34502391

Signed-off-by: Srikanth C S <srikanth.c.s@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Jose Lombera <jose.lombera@oracle.com>