memstate: Print per-process memory usage (RSS) in memory usage summary output
Parsing through /proc/<pid>/smaps_rollup files for all processes to extract
PSS values can get a bit expensive, in terms of CPU usage and time, if done
frequently/periodically. This commit extracts/displays per-process RSS values
instead, in the summary output of memstate. This is much faster, although not
very accurate. If the user wants to know the more accurate PSS numbers, that
can be invoked via the `-p` option. This commit only alters the default summary
output, when memstate is invoked with no options, which is more likely to be
invoked often/periodically.

Signed-off-by: Aruna Ramakrishna <aruna.ramakrishna@oracle.com>
Reviewed-by: Jose Lombera <jose.lombera@oracle.com>