Add vmcore_sz tool subcommand
The vmcore_sz takes the dump_level as an argument and estimates the
vmcore size if a kernel dump is accessed at that moment.  It displays
the total number of pages, pages need to be excluded depending on the
dump level and the expected vmcore size in bytes.  If the dump level is
not specified, the default configured in /etc/kdump.conf will be used.

Orabug: 35824470

Signed-off-by: Partha Satapathy <partha.satapathy@oracle.com>
Co-authored-by: Srikanth C S <srikanth.c.s@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Jose Lombera <jose.lombera@oracle.com>